move utils to v1 folder; tyring to figure out how to make a new Response class
